platform/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
gspread>=5.12.0
google-auth>=2.23.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=10.0.0

